Aggregation of scores is now performed correctly. Parameter definition is also fully accomplished.

+++ src/main/scala/org/apache/spark/ml/feature/MatrixAccumulator.scala
@@ -7,22 +7,22 @@ import org.apache.spark.util.AccumulatorV2
 /**
  * @author sramirez
  */
-class MatrixAccumulator(val rows: Int, val cols: Int) extends AccumulatorV2[Matrix[Double], Matrix[Double]] {
+class MatrixAccumulator(val rows: Int, val cols: Int) extends AccumulatorV2[Matrix[Float], Matrix[Float]] {
   
-  def this(m: Matrix[Double]) = {
+  def this(m: Matrix[Float]) = {
     this(m.rows, m.cols)
     this.accMatrix = m.copy
   }
 
-  private var accMatrix: Matrix[Double] = Matrix.zeros[Double](rows, cols)
+  private var accMatrix: Matrix[Float] = Matrix.zeros[Float](rows, cols)
   private var zero: Boolean = true
 
   def reset(): Unit = {
-    accMatrix = Matrix.zeros[Double](rows, cols)
+    accMatrix = Matrix.zeros[Float](rows, cols)
     zero = true
   }
 
-  def add(v: Matrix[Double]): Unit = {
+  def add(v: Matrix[Float]): Unit = {
     if(isZero) 
       zero = false
     accMatrix += v
@@ -30,9 +30,9 @@ class MatrixAccumulator(val rows: Int, val cols: Int) extends AccumulatorV2[Matr
   
   def isZero(): Boolean = zero
   
-  def merge(other: AccumulatorV2[Matrix[Double], Matrix[Double]]): Unit = accMatrix += other.value
+  def merge(other: AccumulatorV2[Matrix[Float], Matrix[Float]]): Unit = accMatrix += other.value
   
-  def value: Matrix[Double] = accMatrix
+  def value: Matrix[Float] = accMatrix
   
-  def copy(): AccumulatorV2[Matrix[Double], Matrix[Double]] = new MatrixAccumulator(accMatrix)
+  def copy(): AccumulatorV2[Matrix[Float], Matrix[Float]] = new MatrixAccumulator(accMatrix)
 }
